refactor: remove redundant configuration protocols and consolidate taxonomy definitions into taxonomy_core_vo
