+++ modules/cli/src/surface_cli_init_command.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from modules.shared.src.contract_workspace_protocol import IWorkspaceProtocol
+from modules.shared.src.contract_core_protocol import IWorkspaceProtocol
 from modules.shared.src.taxonomy_core_vo import FilePath
 from modules.shared.src.utility_core_response import safe_handle, success_response
