+++ modules/cli/src/surface_cli_run_command.py
@@ -13,7 +13,7 @@ from modules.shared.src.contract_core_aggregate import (
     IDirectPromptAggregate,
     IPromptFileAggregate,
 )
-from modules.shared.src.taxonomy_config_vo import AppConfig
+from modules.shared.src.taxonomy_core_vo import AppConfig
 from modules.shared.src.taxonomy_core_vo import HeadlessFlag
 from modules.shared.src.utility_core_response import error_response, safe_handle, success_response
