+++ modules/core/src/agent_prompt_file_orchestrator.py
@@ -26,7 +26,7 @@ from modules.shared.src.contract_core_protocol import (
     ISendProtocol,
     IStreamProtocol,
 )
-from modules.shared.src.taxonomy_config_vo import AppConfig
+from modules.shared.src.taxonomy_core_vo import AppConfig
 from modules.shared.src.taxonomy_core_error import (
     ResponseDetectionTimeoutError,
 )
