+++ modules/core/src/agent_session_orchestrator.py
@@ -14,7 +14,7 @@ from modules.shared.src.contract_core_protocol import (
     IBrowserProtocol,
     IObservabilityProtocol,
 )
-from modules.shared.src.taxonomy_config_vo import AppConfig
+from modules.shared.src.taxonomy_core_vo import AppConfig
 from modules.shared.src.taxonomy_core_constant import (
     DEFAULT_OUTPUT,
     DEFAULT_TODO,
