+++ modules/core/src/agent_setup_orchestrator.py
@@ -15,7 +15,7 @@ from modules.shared.src.contract_core_protocol import (
     IBrowserProtocol,
     IObservabilityProtocol,
 )
-from modules.shared.src.taxonomy_config_vo import AppConfig
+from modules.shared.src.taxonomy_core_vo import AppConfig
 from modules.shared.src.taxonomy_core_constant import (
     CHAT_URL,
     DEFAULT_OUTPUT,
