+++ modules/core/src/capabilities_prompt_injector.py
@@ -13,7 +13,7 @@ from playwright.sync_api import Error, Page
 from modules.core.src.utility_core_dom_helper import first_visible_element_handle
 from modules.core.src.utility_core_logger_factory import get_logger
 from modules.shared.src.contract_core_protocol import IInjectionProtocol
-from modules.shared.src.taxonomy_config_vo import DEFAULT_INJECTOR_CONFIG, InjectorConfig
+from modules.shared.src.taxonomy_core_vo import DEFAULT_INJECTOR_CONFIG, InjectorConfig
 from modules.shared.src.taxonomy_core_error import ElementNotFoundError, PromptInjectionError
 
 log = get_logger("capabilities_prompt_injector")
