+++ modules/core/src/capabilities_stream_monitor.py
@@ -13,7 +13,7 @@ from modules.core.src.utility_core_dom_helper import is_any_visible, is_selector
 from modules.core.src.utility_core_dom_query import latest_message_text as _dom_latest
 from modules.core.src.utility_core_logger_factory import get_logger
 from modules.shared.src.contract_core_protocol import IStreamProtocol
-from modules.shared.src.taxonomy_config_vo import StreamerConfig
+from modules.shared.src.taxonomy_core_vo import StreamerConfig
 from modules.shared.src.taxonomy_core_constant import (
     SEND_DISABLED_SELECTORS,
     STOP_BUTTON_SELECTORS,
