+++ modules/shared/src/utility_core_path.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 from pathlib import Path
 
-from modules.shared.src.taxonomy_config_vo import AppConfig
+from modules.shared.src.taxonomy_core_vo import AppConfig
 
 SKIP_DIRS: frozenset[str] = frozenset({"done", "failed", ".processing", "proc"})
 ROLE_PATH_SKIP_DIRS: frozenset[str] = frozenset({"todo", "done", "failed", ".processing", "proc"})
