+++ tests/unit_capability_send_dispatcher.py
@@ -10,7 +10,7 @@ from playwright.sync_api import Error
 from modules.core.src.capabilities_send_dispatcher import SendDispatcher
 from modules.core.src.utility_core_dom_query import count_messages, latest_message_text
 from modules.shared.src import LifecycleEmitter, SendDispatchError
-from modules.shared.src.taxonomy_config_vo import SenderConfig
+from modules.shared.src.taxonomy_core_vo import SenderConfig
 from modules.shared.src.taxonomy_core_vo import ClickTimeoutMs
